ref(replay): Move ReplayTable dropdown filters into components/
